# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MS Elevate LEAPS Tracker API
  version: 1.0.0
  description: |
    Complete API documentation for the Microsoft Elevate Indonesia LEAPS Tracker platform.
    
    The LEAPS framework consists of 5 stages:
    - **Learn**: Complete Microsoft Learn courses (20 points)
    - **Explore**: Apply AI in classroom with evidence (50 points)  
    - **Amplify**: Train peers/students (2 points/peer, 1 point/student)
    - **Present**: LinkedIn post with screenshot (20 points)
    - **Shine**: Innovation ideas for recognition
    
    ## Authentication
    All authenticated endpoints require a valid Clerk JWT token in the Authorization header.
    
    ## Rate Limits
    - File uploads: 10MB max per file
    - Amplify submissions: 50 peers, 200 students per 7-day rolling window
    - Bulk operations: Maximum 50 items per request
    
  contact:
    name: MS Elevate Team
    url: https://leaps.mereka.org
servers:
  - url: https://leaps.mereka.org/api
    description: Production server
  - url: https://staging-leaps.mereka.org/api 
    description: Staging server
security:
  - ClerkJWT: []
components:
  securitySchemes:
    ClerkJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk authentication JWT token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Validation error details (for Zod errors)
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: string
        handle:
          type: string
        name:
          type: string
        email:
          type: string
        avatar_url:
          type: string
        school:
          type: string
        cohort:
          type: string
        role:
          type: string
          enum: [participant, reviewer, admin, superadmin]
        profile_visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - handle
        - name
        - email
...
    Submission:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        activity_code:
          type: string
          enum: [LEARN, EXPLORE, AMPLIFY, PRESENT, SHINE]
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        payload:
          type: object
          description: Activity-specific data
        attachments:
          type: array
          items:
            type: string
        reviewer_id:
          type: string
          nullable: true
        review_note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - activity_code
        - status
        - payload
    LearnPayload:
      type: object
      properties:
        provider:
          type: string
          enum: [SPL, ILS]
        course:
          type: string
          minLength: 2
        completedAt:
          type: string
          format: date-time
        certificateFile:
          type: string
          description: Storage path after upload
      required:
        - provider
        - course
        - completedAt
    ExplorePayload:
      type: object
      properties:
        reflection:
          type: string
          minLength: 150
        classDate:
          type: string
          format: date
        school:
          type: string
        evidenceFiles:
          type: array
          items:
            type: string
      required:
        - reflection
        - classDate
    AmplifyPayload:
      type: object
      properties:
        peersTrained:
          type: integer
          minimum: 0
          maximum: 50
        studentsTrained:
          type: integer
          minimum: 0
          maximum: 200
        attendanceProofFiles:
          type: array
          items:
            type: string
      required:
        - peersTrained
        - studentsTrained
    PresentPayload:
      type: object
      properties:
        linkedinUrl:
          type: string
          format: uri
        screenshotFile:
          type: string
        caption:
          type: string
          minLength: 10
      required:
        - linkedinUrl
        - caption
    ShinePayload:
      type: object
      properties:
        ideaTitle:
          type: string
          minLength: 4
        ideaSummary:
          type: string
          minLength: 50
        attachment:
          type: array
          items:
            type: string
      required:
        - ideaTitle
        - ideaSummary
    Badge:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        icon_url:
          type: string
        criteria:
          type: object
        created_at:
          type: string
          format: date-time
      required:
        - id
        - code
        - name
    EarnedBadge:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        badge_id:
          type: string
        earned_at:
          type: string
          format: date-time
        badge:
          $ref: '#/components/schemas/Badge'
      required:
        - id
        - user_id
        - badge_id
        - earned_at
...
paths:
...
  /profile/{handle}:
    get:
      summary: Get public user profile by handle
      tags: [Public]
      security: []
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  publicSubmissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Submission'
                  totalPoints:
                    type: integer
                  earnedBadges:
                    type: array
                    items:
                      $ref: '#/components/schemas/EarnedBadge'
        '404':
          description: User not found or profile is private
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Submissions
  /submissions:
    get:
      summary: Get current user's submissions
      tags: [Submissions]
      parameters:
        - name: activity
          in: query
          schema:
            type: string
            enum: [LEARN, EXPLORE, AMPLIFY, PRESENT, SHINE]
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
      responses:
        '200':
          description: List of user submissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Submission'
                        - type: object
                          properties:
                            activityName:
                              type: string
                            attachmentCount:
                              type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new submission
      tags: [Submissions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                activityCode:
                  type: string
                  enum: [LEARN, EXPLORE, AMPLIFY, PRESENT, SHINE]
                payload:
                  oneOf:
                    - $ref: '#/components/schemas/LearnPayload'
                    - $ref: '#/components/schemas/ExplorePayload'
                    - $ref: '#/components/schemas/AmplifyPayload'
                    - $ref: '#/components/schemas/PresentPayload'
                    - $ref: '#/components/schemas/ShinePayload'
                attachments:
                  type: array
                  items:
                    type: string
                visibility:
                  type: string
                  enum: [PUBLIC, PRIVATE]
                  default: PRIVATE
              required:
                - activityCode
                - payload
      responses:
        '201':
          description: Submission created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      activityCode:
                        type: string
                      status:
                        type: string
                      visibility:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
                      potentialPoints:
                        type: integer
        '400':
          description: Invalid submission data or limits exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Leaderboard
...
  /webhooks/clerk:
    post:
      summary: Clerk user webhook
      tags: [Webhooks]
      security: []
      description: Handles user creation/updates from Clerk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Clerk webhook payload
      responses:
        '200':
          description: Webhook processed successfully
        '400':
          description: Invalid webhook payload
...